Add value type parameter default arguments
Added support for the default value type arguments. This worked
pretty much right away except that the literals might not have
been visited for default value type arguments which is handled
in the expr pass. This is really a hack around this as this
should be handled in a pass prior to the expr pass.