Implemented and tested value parameter passing on Aarch64
Implemented and tested value parameter passing on Aarch64 and
MacOS.

Did an overhaul on the value parameter passing code so that checking
if lowering is needed for parameters is no longer needed in the
client code.

Changed exit  codes on value parameter passing tests as some OS
only support 8-bit exit codes.