Fixed passed by value bugs
Fixed bug with applying pass by value attributes like byval where
variadic parameters FFI functions caused out of bounds access of
ffi_decl->params.

Fixed bug in gen_param where a TK_TUPLETYPE caused an error in
gen_name because members could TK_TYPEPARAMREF. Changed the code
so that reach_type is not called unless it is passbyvalue
annotation. TK_TUPLETYPE should not be used with passbyvalue
anyway. It is possible that this doesn't fix the underlying
problem and a deferred_reify is required in gen_name when looping
through members.

Renamed test x_of_operator to x-of-operator to follow name
convention.