Add nhb capability
Added a capability called nhb. nhb should be used where the
compiler cannot infer that a mutable class or struct can be used
safely in several actors. For example a class that uses atomic
operations in order to ensure that it can be safely shared among
actors. nhb is like a ref but is sendable.