Turn off "arm64 windows partial support" (#4688)
It wasn't actually working as expected. It was cross compiling a
x64 version and then the emulator in Windows 11 was running
everything emulated.

There's some porting that will need to be done to support.

We need to get the arch correct in make.ps1 and the host value used
as well.

We need to add support for `_M_ARM64` in all the places we currently
support `_M_AMD64`/`_M_X64` and possibly other changes. Those
are the ones I have found so far.