Increase full program runner timeout on Intel Mac
The tests are much slower on average than on Linux. This will allow
for more time for "oops" to happen. For example, the regression-1118
test regularly comes in around 45 seconds. A little OS doing something
else and that can timeout before completion.

This commit increases the timeout for full program runner tests from
the default (currently 60 seconds) to 180 seconds.