Add aarch64 linux musl nightly builds (#4706)
This is an exploratory addition that will help shake out any issues
we might have with doing aarch64 builds that get uploaded to
Cloudsmith.

We've never done in a Linux aarch64 environment so it is possible this
might expose issues that we need to address before we start to provide
aarch64 builds and announce them to the world.