Add CTFE Pointer support
Added the support for the Pointer type. This in order to enable
that the classes Array and String works in CTFE.

Made _alloc, _realloc, _apply, _update in the Pointer type public.

Changed the interface for CtfeValue run_method so that the
receiver is passed as a non-const reference.

Added CtfeValueTypeRef in order to handle type references in CTFE.