Moved CTFE to reach pass
Moved the CTFE evaulation to the reach pass. This is the same
location as in ponyta and it also simplifies the evaluation as
the AST has been properly processed in all stages.

It is also benefitial as the type arguments are available from
the calling location which means that type arguments from
class/method outside the comptime scope can be used.