Removed CtfeValueStringLiteral
Removed the CtfeValueStringLiteral class because it is no longer
needed. Instead string literals are converted to the String class
similar to the generated code. In this case it is an allocated
class rather than a immutable global but should be enough to be
able to run the String class code directly.

Added TypeRef information to CtfeValueStruct.