Add CTFE match expression support
Added match support in CTFE. Because of the match expression
relies heavily on the ast type, the custom CtfeValue type was
removed and now relies entirely on the ast type.

Removed CtfeValueTypeRef as the ast type is now being used.

Added initial support for the is expression.

Added recover expression suppoert.