Add CTFE constant objects
Added support for structs and classes to become constant objects.
This means that a ready made version will be stored as constant
data in the executable and be used directly if readonly or
copied if readwrite.

Added flag in order to exploit more CPU cores when building LLVM
on Windows.