Removed the 129-bit literal with signed bit.
Removed the sign bool in lexint_t because since CTFE was moved to
the reach pass, all the literals have been coerced and typed.
This means that calculations on untyped literals are not longer
needed.

Also negative literals are no longer needed as they are all
positive from the beginning. They might be negative when they are
created in CTFE but then it is saved in the bit pattern of the
type.

Changed the names of CtfeValueTypedInt and CtfeValueTypedFloat to
CtfeValueInt and CtfeValueFloat.

Removed CtfeValueIntLiteral as it is no longer needed.