Add CTFE timeout and max recursion
Added a max duration for the for each CTFE execution that can
also be overidden by the command line parameter
--ctfe-max-duration.
Added a max recursion value to the CTFE execution which can be
overridden but the command line parameter --ctfe-max-recursion.

Added the possiblity create constant objects for primitives. Now
returning None or any other primitive from functions should be
possible.

Removed the depth parameter in CTFE evaluations as it is not used
for anything useful.