Move type methods to callback solution
Moved the type specific methods in CtfeValueInt, CtfeValueBool,
CtfeValueFloat and CtfeValuePointer to a std::map instead of
long if..else if chains.
This in order to reduce the boiler plate code and repeating
copy & paste which should make the code more maintainable.

There are remaining calls outside std::map but these are the odd
methods that can't take advantage of repeating code structures.