Add support for primitives, tuples, unions in CTFE arrays
Added support for primitives, tuples and unions in CTFE array.
The values are stored in the array as CtfeValue rather than
the memory layout because it is unlikely that it is interesting
to directly access the raw data in the array. This unlike the
machine word types where conversion between the array happens,
like in the String class for example.